+++ lib/lakos.dart
@@ -2,11 +2,12 @@
 ///
 /// Use [buildModel] to construct a [Model] object.
 ///
-/// Use [convertModelToDirectedGraph] for further analysis with the directed_graph library.
+/// Use [Model.getOutput] to print the model in dot or JSON format.
 ///
-/// See example/example.dart.
+/// Use [Model.toDirectedGraph] for further analysis with the `directed_graph` library.
+///
+/// See [example/example.dart](https://github.com/OlegAlexander/lakos/blob/master/example/example.dart).
 library lakos;
 
 export 'src/build_model.dart' show buildModel, PubspecYamlNotFoundException;
-export 'src/metrics.dart' show convertModelToDirectedGraph;
 export 'src/model.dart';